Comment out vec<str> template calls to hdf writers
This is a total hack, but at least allows bam2bax to complete without
segfaulting.  This tool is not really supported any more so I don't want
to spend any more time trying to fix the real problem(s).  At least I've
isolated the problem area a bit more so if we do want to fix it
eventually, we'll have a better place to start.